added new categories
