+++ ersilia/hub/content/card.py
@@ -2,43 +2,12 @@ import os
 import json
 import requests
 import yaml
+from .base_information import BaseInformation
 from ... import ErsiliaBase
 from ...utils.terminal import run_command
 from ...auth.auth import Auth
 from ...db.hubdata.interfaces import AirtableInterface
 from ...db.hubdata.json_models_interface import JsonModelsInterface
-import validators
-
-try:
-    from validators import ValidationFailure
-except ImportError:
-    from validators import ValidationError as ValidationFailure
-
-from ...utils.exceptions_utils.card_exceptions import (
-    SlugBaseInformationError,
-    IdentifierBaseInformationError,
-    StatusBaseInformationError,
-    TitleBaseInformationError,
-    DescriptionBaseInformationError,
-    ModeBaseInformationError,
-    InputBaseInformationError,
-    InputShapeBaseInformationError,
-    OutputBaseInformationError,
-    OutputTypeBaseInformationError,
-    OutputShapeBaseInformationError,
-    TaskBaseInformationError,
-    TagBaseInformationError,
-    PublicationBaseInformationError,
-    SourceCodeBaseInformationError,
-    LicenseBaseInformationError,
-    GithubBaseInformationError,
-    DockerhubBaseInformationError,
-    DockerArchitectureInformationError,
-    S3BaseInformationError,
-    BothIdentifiersBaseInformationError,
-    MemoryGbBaseInformationError,
-)
-from ...utils.identifiers.model import ModelIdentifier
 from ...utils.logging import make_temp_dir
 
 try:
@@ -53,385 +22,6 @@ from ...default import (
     METADATA_YAML_FILE,
 )
 from ...utils.paths import get_metadata_from_base_dir
-
-
-class BaseInformation(ErsiliaBase):
-    def __init__(self, config_json):
-        ErsiliaBase.__init__(self, config_json=config_json, credentials_json=None)
-        self._github = None
-        self._identifier = None
-        self._slug = None
-        self._status = None
-        self._title = None
-        self._description = None
-        self._mode = None
-        self._task = None
-        self._input = None
-        self._input_shape = None
-        self._output = None
-        self._output_type = None
-        self._output_shape = None
-        self._interpretation = None
-        self._tag = None
-        self._publication = None
-        self._source_code = None
-        self._license = None
-        self._contributor = None
-        self._dockerhub = None
-        self._docker_architecture = None
-        self._s3 = None
-        self._memory_gb = None
-
-    def _is_valid_url(self, url_string: str) -> bool:
-        result = validators.url(url_string)
-        if isinstance(result, ValidationFailure):
-            return False
-        return result
-
-    def _read_default_fields(self, field):
-        root = os.path.dirname(os.path.abspath(__file__))
-        filename = field.lower().replace(" ", "_")
-        file_path = os.path.join(root, "metadata", filename + ".txt")
-        with open(file_path, "r") as f:
-            valid_field = f.read().split("\n")
-        return valid_field
-
-    @property
-    def identifier(self):
-        return self._identifier
-
-    @identifier.setter
-    def identifier(self, new_identifier):
-        mi = ModelIdentifier()
-        if not mi.is_valid(new_identifier):
-            raise IdentifierBaseInformationError
-        self._identifier = new_identifier
-
-    @property
-    def slug(self):
-        return self._slug
-
-    @slug.setter
-    def slug(self, new_slug):
-        if new_slug.lower() != new_slug:
-            raise SlugBaseInformationError
-        if len(new_slug) > 60:
-            raise SlugBaseInformationError
-        if len(new_slug) < 5:
-            raise SlugBaseInformationError
-        self._slug = new_slug
-
-    @property
-    def status(self):
-        return self._status
-
-    @status.setter
-    def status(self, new_status):
-        if new_status not in self._read_default_fields("Status"):
-            raise StatusBaseInformationError
-        self._status = new_status
-
-    @property
-    def title(self):
-        return self._title
-
-    @title.setter
-    def title(self, new_title):
-        if len(new_title) > 300:
-            raise TitleBaseInformationError
-        if len(new_title) < 10:
-            raise TitleBaseInformationError
-        self._title = new_title
-
-    @property
-    def description(self):
-        return self._description
-
-    @description.setter
-    def description(self, new_description):
-        if len(new_description) < 200:
-            raise DescriptionBaseInformationError
-        if new_description == self._title:
-            raise DescriptionBaseInformationError
-        self._description = new_description
-
-    @property
-    def mode(self):
-        return self._mode
-
-    @mode.setter
-    def mode(self, new_mode):
-        if new_mode not in self._read_default_fields("Mode"):
-            raise ModeBaseInformationError
-        self._mode = new_mode
-
-    @property
-    def input(self):
-        return self._input
-
-    @input.setter
-    def input(self, new_input):
-        if type(new_input) is str:
-            new_input = [new_input]
-        if type(new_input) is not list:
-            raise InputBaseInformationError
-        for inp in new_input:
-            if inp not in self._read_default_fields("Input"):
-                raise InputBaseInformationError
-        self._input = new_input
-
-    @property
-    def input_shape(self):
-        return self._input_shape
-
-    @input_shape.setter
-    def input_shape(self, new_input_shape):
-        if new_input_shape not in self._read_default_fields("Input Shape"):
-            raise InputShapeBaseInformationError
-        self._input_shape = new_input_shape
-
-    @property
-    def task(self):
-        return self._task
-
-    @task.setter
-    def task(self, new_task):
-        if type(new_task) is str:
-            new_task = [new_task]
-        if type(new_task) is not list:
-            raise TaskBaseInformationError
-        for nt in new_task:
-            if nt not in self._read_default_fields("Task"):
-                raise TaskBaseInformationError
-        self._task = new_task
-
-    @property
-    def output(self):
-        return self._output
-
-    @output.setter
-    def output(self, new_output):
-        if type(new_output) is str:
-            new_output = [new_output]
-        default_output = self._read_default_fields("Output")
-        for no in new_output:
-            if no not in default_output:
-                raise OutputBaseInformationError
-        self._output = new_output
-
-    @property
-    def output_type(self):
-        return self._output_type
-
-    @output_type.setter
-    def output_type(self, new_output_type):
-        if type(new_output_type) is str:
-            new_output_type = [new_output_type]
-        default_output_type = self._read_default_fields("Output Type")
-        for no in new_output_type:
-            if no not in default_output_type:
-                raise OutputTypeBaseInformationError
-        self._output_type = new_output_type
-
-    @property
-    def output_shape(self):
-        return self._output_shape
-
-    @output_shape.setter
-    def output_shape(self, new_output_shape):
-        default_output_shape = self._read_default_fields("Output Shape")
-        if new_output_shape not in default_output_shape:
-            raise OutputShapeBaseInformationError
-        self._output_shape = new_output_shape
-
-    @property
-    def interpretation(self):
-        return self._interpretation
-
-    @interpretation.setter
-    def interpretation(self, new_interpretation):
-        self._interpretation = new_interpretation
-
-    @property
-    def tag(self):
-        return self._tag
-
-    @tag.setter
-    def tag(self, new_tag):
-        if type(new_tag) is str:
-            new_tag = [new_tag]
-        if type(new_tag) is not list:
-            raise TagBaseInformationError
-        default_tags = self._read_default_fields("Tag")
-        for nt in new_tag:
-            if nt not in default_tags:
-                raise TagBaseInformationError
-        self._tag = new_tag
-
-    @property
-    def publication(self):
-        return self._publication
-
-    @publication.setter
-    def publication(self, new_publication):
-        if not self._is_valid_url(new_publication):
-            raise PublicationBaseInformationError
-        self._publication = new_publication
-
-    @property
-    def source_code(self):
-        return self._source_code
-
-    @source_code.setter
-    def source_code(self, new_source_code):
-        if not self._is_valid_url(new_source_code):
-            raise SourceCodeBaseInformationError
-        self._source_code = new_source_code
-
-    @property
-    def license(self):
-        return self._license
-
-    @license.setter
-    def license(self, new_license):
-        if new_license not in self._read_default_fields("License"):
-            raise LicenseBaseInformationError
-        self._license = new_license
-
-    @property
-    def date(self):
-        return self._date
-
-    @date.setter
-    def date(self, new_date):
-        self._date = new_date
-
-    @property
-    def contributor(self):
-        return self._contributor
-
-    @contributor.setter
-    def contributor(self, new_contributor):
-        self._contributor = new_contributor
-
-    @property
-    def github(self):
-        model_id = self.identifier
-        if model_id is None:
-            raise GithubBaseInformationError
-        self._github = "https://github.com/ersilia-os/{0}".format(model_id)
-        return self._github
-
-    @property
-    def dockerhub(self):
-        return self._dockerhub
-
-    @dockerhub.setter
-    def dockerhub(self, new_dockerhub_url):
-        if not new_dockerhub_url.startswith("https://hub.docker.com/r/ersiliaos/"):
-            raise DockerhubBaseInformationError
-        self._dockerhub = new_dockerhub_url
-
-    @property
-    def docker_architecture(self):
-        return self._docker_architecture
-
-    @docker_architecture.setter
-    def docker_architecture(self, new_docker_architecture):
-        if type(new_docker_architecture) is str:
-            new_docker_architecture = [new_docker_architecture]
-        for d in new_docker_architecture:
-            if d not in self._read_default_fields("Docker Architecture"):
-                raise DockerArchitectureInformationError
-        self._docker_architecture = new_docker_architecture
-
-    @property
-    def s3(self):
-        return self._s3
-
-    @s3.setter
-    def s3(self, new_s3_url):
-        if not new_s3_url.startswith(
-            "https://ersilia-models-zipped.s3.eu-central-1.amazonaws.com/"
-        ):
-            raise S3BaseInformationError
-        self._s3 = new_s3_url
-
-    @property
-    def both_identifiers(self):
-        model_id = self.identifier
-        slug = self.slug
-        if model_id is None or slug is None:
-            raise BothIdentifiersBaseInformationError
-        self._both_identifiers = (model_id, slug)
-        return self._both_identifiers
-
-    @property
-    def memory_gb(self):
-        return self._memory_gb
-
-    @memory_gb.setter
-    def memory_gb(self, new_memory_gb):
-        if type(new_memory_gb) != int:
-            raise MemoryGbBaseInformationError
-        self._memory_gb = new_memory_gb
-
-    def as_dict(self):
-        data = {
-            "Identifier": self.identifier,
-            "Slug": self.slug,
-            "Status": self.status,
-            "Title": self.title,
-            "Description": self.description,
-            "Mode": self.mode,
-            "Input": self.input,
-            "Input Shape": self.input_shape,
-            "Task": self.task,
-            "Output": self.output,
-            "Output Type": self.output_type,
-            "Output Shape": self.output_shape,
-            "Interpretation": self.interpretation,
-            "Tag": self.tag,
-            "Publication": self.publication,
-            "Source Code": self.source_code,
-            "License": self.license,
-            "Contributor": self.contributor,
-            "DockerHub": self.dockerhub,
-            "Docker Architecture": self.docker_architecture,
-            "S3": self.s3,
-            "Memory Gb": self.memory_gb,
-        }
-        data = dict((k, v) for k, v in data.items() if v is not None)
-        return data
-
-    def from_dict(self, data):
-        self.identifier = data["Identifier"]
-        self.slug = data["Slug"]
-        self.status = data["Status"]
-        self.title = data["Title"]
-        self.description = data["Description"]
-        self.mode = data["Mode"]
-        self.input = data["Input"]
-        self.input_shape = data["Input Shape"]
-        self.task = data["Task"]
-        self.output = data["Output"]
-        self.output_type = data["Output Type"]
-        self.output_shape = data["Output Shape"]
-        self.interpretation = data["Interpretation"]
-        self.tag = data["Tag"]
-        self.publication = data["Publication"]
-        self.source_code = data["Source Code"]
-        self.license = data["License"]
-        if "Contributor" in data:
-            self.contributor = data["Contributor"]
-        if "DockerHub" in data:
-            self.dockerhub = data["DockerHub"]
-        if "Docker Architecture" in data:
-            self.docker_architecture = data["Docker Architecture"]
-        if "S3" in data:
-            self.s3 = data["S3"]
-        if "Memory Gb" in data:
-            self.memory_gb = data["Memory Gb"]
 
 
 class RepoMetadataFile(ErsiliaBase):
@@ -779,6 +369,7 @@ class LocalCard(ErsiliaBase):
         else:
             return
 
+
 class LakeCard(ErsiliaBase):
     def __init__(self, config_json=None):
         ErsiliaBase.__init__(self, config_json=config_json)
@@ -820,6 +411,7 @@ class S3JsonCard(JsonModelsInterface):
             return self.get_card_by_slug(slug)
         else:
             raise ValueError("Either model_id or slug must be provided")
+
 
 class ModelCard(object):
     def __init__(self, config_json=None):
