+++ ersilia/hub/content/columns/data_types.txt
@@ -1,0 +1,4 @@
+float
+integer
+string
+other
