+++ ersilia/hub/content/columns/desired_directions.txt
@@ -1,0 +1,3 @@
+high
+intermediate
+low
