+++ ersilia/hub/content/metadata/biomedical_area.txt
@@ -1,0 +1,3 @@
+ADMET
+Malaria
+Tuberculosis
