+++ ersilia/hub/content/metadata/output.txt
@@ -1,5 +1,7 @@
-Boolean
 Compound
+Score
+Value
+Boolean
 Descriptor
 Distance
 Experimental value
@@ -7,5 +9,4 @@ Image
 Other value
 Probability
 Protein
-Score
 Text
