+++ ersilia/hub/content/metadata/publication_type.txt
@@ -1,0 +1,3 @@
+Peer reviewed
+Preprint
+Other
