+++ ersilia/hub/content/metadata/source_type.txt
@@ -1,0 +1,3 @@
+External
+Replicated
+Internal
