+++ ersilia/hub/content/metadata/subtask.txt
@@ -1,0 +1,6 @@
+Featurization
+Projection
+Property calculation or prediction
+Activity prediction
+Similarity search
+Generation
