+++ ersilia/hub/content/metadata/target_organism.txt
@@ -1,0 +1,3 @@
+Human
+Plasmodium spp.
+Mycobacterium tuberculosis
