+++ ersilia/hub/content/metadata/task.txt
@@ -1,7 +1,9 @@
+Representation
+Annotation
+Sampling
 Classification
 Regression
 Generative
-Representation
 Similarity
 Clustering
 Dimensionality reduction
